+++ lib/jazzy/sourcekitten.rb
@@ -223,24 +223,13 @@ module Jazzy
         doc['key.parsed_declaration'] || doc['key.doc.declaration'],
         Config.instance.objc_mode ? 'objc' : 'swift',
       )
-      declaration.abstract = comment_from_doc(doc)
+      declaration.abstract = Jazzy.markdown.render(doc['key.doc.comment'] || '')
       declaration.discussion = ''
       declaration.return = make_paragraphs(doc, 'key.doc.result_discussion')
 
       declaration.parameters = parameters(doc)
 
       @documented_count += 1
-    end
-
-    def self.comment_from_doc(doc)
-      swift_version = Config.instance.swift_version.to_f
-      comment = doc['key.doc.comment'] || ''
-      if swift_version < 2
-        # comment until first ReST definition
-        matches = /^\s*:[^\s]+:/.match(comment)
-        comment = comment[0...matches.begin(0)] if matches
-      end
-      Jazzy.markdown.render(comment)
     end
 
     def self.make_substructure(doc, declaration)