+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## Master
+
+##### Breaking
+
+* Config files now use the same option names as the command line. If you are
+  using one of the keys that has changed in your `.jazzy.yaml`, you will receive
+  a warning. See the [pull request](https://github.com/realm/jazzy/pull/456) for
+  a complete list of changed options. As always, you can get a list of all
+  options with `jazzy --help config`.
+  [pcantrell](https://github.com/pcantrell)
+
 ## 0.5.0
 
 ##### Breaking